refactor: separate ROS2-free EKF core
